Feature: Updates Scenes templates
