--- packages/create-plugin/templates/scenes-app/src/pages/HelloWorld/HelloWorld.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { getScene } from './helloWorldScene';
-
-const HelloWorld = () => {
-  const scene = getScene();
-
-  return <scene.Component model={scene} />;
-};
-
-export default HelloWorld;
