+++ packages/create-plugin/templates/scenes-app/src/pages/HelloWorld/helloWorldPage.ts
@@ -1,0 +1,10 @@
+import { SceneAppPage } from '@grafana/scenes';
+import { helloWorldScene } from './helloWorldScene';
+import { prefixRoute } from '../../utils/utils.routing';
+import { ROUTES } from '../../constants';
+
+export const helloWorldPage = new SceneAppPage({
+  title: 'Hello World',
+  url: prefixRoute(ROUTES.HelloWorld),
+  getScene: helloWorldScene,
+});
