+++ packages/create-plugin/templates/scenes-app/src/pages/HelloWorld/helloWorldScene.ts
@@ -1,6 +1,6 @@
 import { EmbeddedScene, SceneFlexLayout, SceneFlexItem, PanelBuilders } from '@grafana/scenes';
 
-export function getScene() {
+export function helloWorldScene() {
   return new EmbeddedScene({
     body: new SceneFlexLayout({
       children: [
