+++ packages/create-plugin/templates/scenes-app/src/pages/Home/homePage.ts
@@ -1,0 +1,12 @@
+import { SceneAppPage } from '@grafana/scenes';
+import { homeScene } from './homeScene';
+import { prefixRoute } from '../../utils/utils.routing';
+import { ROUTES } from '../../constants';
+
+export const homePage = new SceneAppPage({
+  title: 'Home page',
+  url: prefixRoute(ROUTES.Home),
+  subTitle:
+    'This scene showcases a basic scene functionality, including query runner, variable and a custom scene object.',
+  getScene: () => homeScene(),
+});
