+++ packages/create-plugin/templates/scenes-app/src/pages/WithDrilldown/temperatureOverviewScene.ts
@@ -1,19 +1,19 @@
 import { ReducerID } from '@grafana/data';
 import {
   EmbeddedScene,
+  PanelBuilders,
+  SceneControlsSpacer,
+  SceneFlexItem,
   SceneFlexLayout,
   SceneQueryRunner,
-  SceneTimeRange,
-  SceneTimePicker,
-  SceneControlsSpacer,
   SceneRefreshPicker,
-  SceneFlexItem,
-  PanelBuilders,
+  SceneTimePicker,
+  SceneTimeRange,
 } from '@grafana/scenes';
 import { DATASOURCE_REF } from '../../constants';
-import { getRoomTemperatureStatPanel } from './panels';
+import { RoomTemperatureStat } from './RoomTemperatureStat';
 
-export function getTemperatureOverviewScene(roomName: string) {
+export function temperatureOverviewScene(roomName: string) {
   return new EmbeddedScene({
     $timeRange: new SceneTimeRange({
       from: 'now-12h',
@@ -36,40 +36,16 @@ export function getTemperatureOverviewScene(roomName: string) {
             direction: 'row',
             children: [
               new SceneFlexItem({
-                body: getRoomTemperatureStatPanel([ReducerID.min]),
+                body: RoomTemperatureStat([ReducerID.min]),
               }),
               new SceneFlexItem({
-                body: getRoomTemperatureStatPanel([ReducerID.max]),
+                body: RoomTemperatureStat([ReducerID.max]),
               }),
               new SceneFlexItem({
-                body: getRoomTemperatureStatPanel([ReducerID.mean]),
+                body: RoomTemperatureStat([ReducerID.mean]),
               }),
             ],
           }),
-        }),
-      ],
-    }),
-    controls: [new SceneControlsSpacer(), new SceneTimePicker({ isOnCanvas: true }), new SceneRefreshPicker({})],
-  });
-}
-
-export function getHumidityOverviewScene(roomName: string) {
-  return new EmbeddedScene({
-    $timeRange: new SceneTimeRange({
-      from: 'now-12h',
-      to: 'now',
-    }),
-    $data: new SceneQueryRunner({
-      datasource: DATASOURCE_REF,
-      queries: [getRoomHumidityQuery(roomName)],
-      maxDataPoints: 100,
-    }),
-    body: new SceneFlexLayout({
-      direction: 'column',
-      children: [
-        new SceneFlexItem({
-          height: 500,
-          body: PanelBuilders.timeseries().setTitle('Humidity over time').setUnit('humidity').build(),
         }),
       ],
     }),
@@ -86,17 +62,5 @@ function getRoomTemperatureQuery(roomName: string) {
     alias: roomName,
     min: 10,
     max: 30,
-  };
-}
-
-function getRoomHumidityQuery(roomName: string) {
-  return {
-    refId: 'Humidity',
-    datasource: DATASOURCE_REF,
-    scenarioId: 'random_walk',
-    seriesCount: 1,
-    alias: roomName,
-    min: 30,
-    max: 60,
   };
 }
